--- code/numerical-integration/Model/src/main/java/ru/unn/agile/numerical_integration/Expression.java
@@ -1,5 +1,0 @@
-package ru.unn.agile.numerical_integration;
-
-public interface Expression {
-    double execute(double x);
-};
