+++ code/numerical-integration/model/src/main/java/ru/unn/agile/numerical_integration/model/Expression.java
@@ -1,0 +1,6 @@
+package ru.unn.agile.numerical_integration.Model;
+
+public interface Expression {
+    double execute(double x);
+}
+
