+++ code/numerical-integration/view/src/main/java/ru/unn/agile/numerical_integration/view/legacy/MainForm.form
