+++ code/numerical-integration/view/src/test/java/ru/unn/agile/numerical_integration/view/legacy/FakeViewTests.java
