+++ code/numerical-integration/viewmodel/src/main/java/ru/unn/agile/numerical_integration/viewmodel/legacy/MathParser.java
