+++ code/numerical-integration/viewmodel/src/test/java/ru/unn/agile/numerical_integration/viewmodel/legacy/MathParserTests.java
