+++ settings.gradle
@@ -11,12 +11,12 @@ include "code/pozdiaev-valerii-currency-converter"
 project(":code/pozdiaev-valerii-currency-converter").name = "pozdiaev-valerii-currency-converter"
 
 // nickshevr, Zhiltsov Max
-include "code/numerical-integration/Model"
-project(":code/numerical-integration/Model").name = "numerical-integration-Model-legacy"
-include "code/numerical-integration/ViewModel"
-project(":code/numerical-integration/ViewModel").name = "numerical-integration-ViewModel-legacy"
-include "code/numerical-integration/View"
-project(":code/numerical-integration/View").name = "numerical-integration-View-legacy"
+include "code/numerical-integration/model"
+project(":code/numerical-integration/model").name = "numerical-integration-Model-legacy"
+include "code/numerical-integration/viewmodel"
+project(":code/numerical-integration/viewmodel").name = "numerical-integration-ViewModel-legacy"
+include "code/numerical-integration/view"
+project(":code/numerical-integration/view").name = "numerical-integration-View-legacy"
 
 // Korniakov Kirill, lab1
 include "code/korniakov-kirill-lab1"
